+++ doc/changes.rst
@@ -2,7 +2,16 @@
 Change Log
 ==========
 
-1.9.7 (unreleased)
+1.9.8 (unreleased)
+------------------
+
+* Adds redrock, surveysim, desisurvey, and healpy to dependency version checks.
+* Adds redrock and surveysim to known products for installation.
+* Fix team name in license file.
+* Support new ``/global/common/software`` filesystem at NERSC.
+* Support ``coriknl`` versions of desiconda_.
+
+1.9.7 (2017-09-29)
 ------------------
 
 * Fixed some test failures that occurred in the NERSC environment and/or
@@ -12,12 +21,15 @@ Change Log
 * Fixed bugs that could cause bricks to be slightly too big, and that
   incorrectly special-cased the north pole with brick sizes that don't
   evenly divide 180 degrees (PR `#84`_).
+* Adds ``return_grid_data``` option to :func:``desiutil.plots.plot_sky_binned``
+  (PR `#85`_).
 * Added tests of :mod:`desiutil.sklearn` (PR `#86`_).
 
 .. _`#80`: https://github.com/desihub/desiutil/pull/80
 .. _`#81`: https://github.com/desihub/desiutil/pull/81
 .. _`#84`: https://github.com/desihub/desiutil/pull/84
 .. _`#86`: https://github.com/desihub/desiutil/pull/86
+.. _`#85`: https://github.com/desihub/desiutil/pull/85
 
 
 1.9.6 (2017-07-12)